[Bronze V] Title: 이상한 기호, Time: 124 ms, Memory: 14184 KB -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -6,10 +6,10 @@ public class Main {
         BufferedWriter bufferedWriter = new BufferedWriter(new OutputStreamWriter(System.out));
 
         String[] input = bufferedReader.readLine().split(" ");
-        int a = Integer.parseInt(input[0]);
-        int b = Integer.parseInt(input[1]);
+        long a = Integer.parseInt(input[0]);
+        long b = Integer.parseInt(input[1]);
 
-        int result = (a+b)*(a-b);
+        long result = (a+b)*(a-b);
         bufferedWriter.write(String.valueOf(result));
         bufferedReader.close();
         bufferedWriter.close();
